Done Profile

+++ app/app/src/main/java/com/fpt/canteenrunner/Database/Model/AccountEntity.java
@@ -16,7 +16,7 @@ public class AccountEntity implements Serializable {
     public String Password;
     public String Email;
     public String PhoneNumber;
-    public double Score;
+    public int Score;
     public String Role;
     public String CreatedDate;
     public boolean IsFingerPrintEnabled;
@@ -25,7 +25,7 @@ public class AccountEntity implements Serializable {
     public AccountEntity() {
     }
 
-    public AccountEntity(@NonNull String accountID, String username, String password, String email, String phoneNumber, double score, String role, String createdDate, boolean isFingerPrintEnabled, String fingerPrintData) {
+    public AccountEntity(@NonNull String accountID, String username, String password, String email, String phoneNumber, int score, String role, String createdDate, boolean isFingerPrintEnabled, String fingerPrintData) {
         AccountID = accountID;
         Username = username;
         Password = password;
@@ -91,7 +91,7 @@ public class AccountEntity implements Serializable {
         return Score;
     }
 
-    public void setScore(double score) {
+    public void setScore(int score) {
         Score = score;
     }
